XMLHttpRequest/tests/submissions/Opera/setrequestheader-bogus-value.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: setRequestHeader() value argument checks</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      function try_value(value) {
        assert_throws("INVALID_STATE_ERR", function() { client.setRequestHeader("x-test", value) })
      }
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "...")
        try_value("t\rt")
        try_value("t\nt")
      })
    </script>
  </body>
</html>
